Seperate tests.py into several test files
