+++ tests/thumbnail_tests/utils.py
@@ -1,11 +1,28 @@
 # coding: utf-8
 from __future__ import unicode_literals
 import os
+import shutil
+import unittest
+import logging
 from contextlib import contextmanager
 from subprocess import check_output
 
+from PIL import Image
 from django.test.signals import setting_changed
 from django.conf import UserSettingsHolder
+
+from sorl.thumbnail.conf import settings
+from sorl.thumbnail.helpers import get_module_class
+from sorl.thumbnail.images import ImageFile
+from sorl.thumbnail.log import ThumbnailLogHandler
+from .models import Item
+from .storage import MockLoggingHandler
+
+DATA_DIR = os.path.join(settings.MEDIA_ROOT, 'data')
+
+handler = ThumbnailLogHandler()
+handler.setLevel(logging.ERROR)
+logging.getLogger('sorl.thumbnail').addHandler(handler)
 
 
 @contextmanager
@@ -65,3 +82,67 @@ class override_custom_settings(object):
             new_value = getattr(self.settings, key, None)
             setting_changed.send(sender=self.settings._wrapped.__class__,
                                  setting=key, value=new_value, enter=False)
+
+
+class FakeFile(object):
+    """
+    Used to test the _get_format method.
+    """
+
+    def __init__(self, name):
+        self.name = name
+
+
+class BaseTestCase(unittest.TestCase):
+    IMAGE_DIMENSIONS = [(500, 500), (100, 100), (200, 100), ]
+    BACKEND = None
+    ENGINE = None
+    KVSTORE = None
+
+    def create_image(self, name, dim):
+        """
+        Creates an image and prepends the MEDIA ROOT path.
+        :param name: e.g. 500x500.jpg
+        :param dim: a dimension tuple e.g. (500, 500)
+        """
+        filename = os.path.join(settings.MEDIA_ROOT, name)
+        im = Image.new('L', dim)
+        im.save(filename)
+        return Item.objects.get_or_create(image=name)
+
+    def setUp(self):
+        self.BACKEND = get_module_class(settings.THUMBNAIL_BACKEND)()
+        self.ENGINE = get_module_class(settings.THUMBNAIL_ENGINE)()
+        self.KVSTORE = get_module_class(settings.THUMBNAIL_KVSTORE)()
+
+        if not os.path.exists(settings.MEDIA_ROOT):
+            os.makedirs(settings.MEDIA_ROOT)
+            shutil.copytree(settings.DATA_ROOT, DATA_DIR)
+
+        for dimension in self.IMAGE_DIMENSIONS:
+            name = '%sx%s.jpg' % dimension
+            self.create_image(name, dimension)
+
+    def tearDown(self):
+        shutil.rmtree(settings.MEDIA_ROOT)
+
+
+class BaseStorageTestCase(unittest.TestCase):
+    image = None
+    name = None
+
+    def setUp(self):
+        os.makedirs(settings.MEDIA_ROOT)
+        filename = os.path.join(settings.MEDIA_ROOT, self.name)
+        Image.new('L', (100, 100)).save(filename)
+        self.image = ImageFile(self.name)
+
+        logger = logging.getLogger('slog')
+        logger.setLevel(logging.DEBUG)
+        handler = MockLoggingHandler(level=logging.DEBUG)
+        logger.addHandler(handler)
+        self.log = handler.messages['debug']
+
+    def tearDown(self):
+        shutil.rmtree(settings.MEDIA_ROOT)
+
